refactor(cli): modularize inference and provider features
